requirements.txt
streamlit
pillow
psycopg2
numpy
smtplib
email
os
json
datetime
